Refactor out the common ConsoleProps class
