+++ src/io/flutter/run/test/TestLaunchState.java
@@ -30,6 +30,7 @@ import com.jetbrains.lang.dart.util.DartUrlResolver;
 import io.flutter.FlutterBundle;
 import io.flutter.console.FlutterConsoleFilter;
 import io.flutter.pub.PubRoot;
+import io.flutter.run.common.ConsoleProps;
 import io.flutter.run.daemon.DaemonConsoleView;
 import io.flutter.run.daemon.RunMode;
 import io.flutter.sdk.FlutterCommandStartResult;
@@ -116,8 +117,8 @@ class TestLaunchState extends CommandLineState {
     // Create a console showing a test tree.
     final Project project = getEnvironment().getProject();
     final DartUrlResolver resolver = DartUrlResolver.getInstance(project, testFileOrDir);
-    final ConsoleProps props = new ConsoleProps(config, executor, resolver);
-    final BaseTestsOutputConsoleView console = SMTestRunnerConnectionUtil.createConsole("FlutterTestRunner", props);
+    final ConsoleProps props = ConsoleProps.forPub(config, executor, resolver);
+    final BaseTestsOutputConsoleView console = SMTestRunnerConnectionUtil.createConsole(ConsoleProps.pubFrameworkName, props);
     final Module module = ModuleUtil.findModuleForFile(testFileOrDir, project);
     if (module != null) {
       console.addMessageFilter(new FlutterConsoleFilter(module));
@@ -150,38 +151,5 @@ class TestLaunchState extends CommandLineState {
     return pubRoot;
   }
 
-  /**
-   * Configuration for the test console.
-   * <p>
-   * In particular, configures how it parses test events and handles the re-run action.
-   */
-  private static class ConsoleProps extends SMTRunnerConsoleProperties implements SMCustomMessagesParsing {
-    @NotNull
-    private final DartUrlResolver resolver;
 
-    public ConsoleProps(@NotNull TestConfig config, @NotNull Executor exec, @NotNull DartUrlResolver resolver) {
-      super(config, "FlutterTestRunner", exec);
-      this.resolver = resolver;
-      setUsePredefinedMessageFilter(false);
-      setIdBasedTestTree(true);
-    }
-
-    @Nullable
-    @Override
-    public SMTestLocator getTestLocator() {
-      return FlutterTestLocationProvider.INSTANCE;
-    }
-
-    @Override
-    public OutputToGeneralTestEventsConverter createTestEventsConverter(@NotNull String testFrameworkName,
-                                                                        @NotNull TestConsoleProperties props) {
-      return new FlutterTestEventsConverter(testFrameworkName, props, resolver);
-    }
-
-    @Nullable
-    @Override
-    public AbstractRerunFailedTestsAction createRerunFailedTestsAction(ConsoleView consoleView) {
-      return null; // TODO(skybrian) implement
-    }
-  }
 }
